Move custom field test and add test with 2 base class
