+++ tests/assert_type/db/models/fields/test_custom_fields.py
@@ -160,3 +160,53 @@ def field_two_typevar_form_in_user_annotation() -> None:
     instance.field = "no"  # type: ignore[call-overload]  # pyrefly: ignore[no-matching-overload]  # ty: ignore[invalid-assignment]  # pyright: ignore[reportAttributeAccessIssue]
     instance.null_field = None  # type: ignore[call-overload]  # pyrefly: ignore[no-matching-overload]  # ty: ignore[invalid-assignment]  # pyright: ignore[reportAttributeAccessIssue]
     instance.null_field = CustomFieldValue()
+
+
+def nullable_subclass_inherits_null_overload() -> None:
+    _ST_Text = TypeVar("_ST_Text", contravariant=True, default=str | int | Combinable)
+    _GT_Text = TypeVar("_GT_Text", covariant=True, default=str)
+    _ST_Int = TypeVar("_ST_Int", contravariant=True, default=float | int | str | Combinable)
+    _GT_Int = TypeVar("_GT_Int", covariant=True, default=int)
+    _NT = TypeVar("_NT", Literal[True], Literal[False], default=Literal[False])
+
+    class HtmlField(models.TextField[_ST_Text, _GT_Text, _NT]): ...
+
+    class IntWrap(models.IntegerField[_ST_Int, _GT_Int, _NT]): ...
+
+    class Article(models.Model):
+        body = HtmlField()
+        body_nullable = HtmlField(null=True)
+        count = IntWrap()
+        count_nullable = IntWrap(null=True)
+
+    assert_type(Article().body, str)
+    assert_type(Article().body_nullable, str | None)
+    assert_type(Article().count, int)
+    assert_type(Article().count_nullable, int | None)
+
+
+def nullable_field_subclass_without_explicit_type_vars() -> None:
+    """
+    We auto add typevars in mypy plugin which avoid issues here.
+
+    TODO: False positive pyrefly/ty/pyright
+    """
+
+    class HTMLField(models.TextField): ...
+
+    class IntWrap(models.IntegerField): ...
+
+    class FieldMixin: ...
+
+    class MySlugField(models.SlugField, FieldMixin): ...
+
+    class Article(models.Model):
+        body = HTMLField()
+        body_nullable = HTMLField(null=True)  # pyright: ignore[reportArgumentType]  # pyrefly: ignore[bad-argument-type]  # ty: ignore[invalid-argument-type]
+        count_nullable = IntWrap(null=True)  # pyright: ignore[reportArgumentType]  # pyrefly: ignore[bad-argument-type]  # ty: ignore[invalid-argument-type]
+        slug_nullable = MySlugField(null=True)  # pyright: ignore[reportArgumentType]  # pyrefly: ignore[bad-argument-type]  # ty: ignore[invalid-argument-type]
+
+    assert_type(Article().body, str)
+    assert_type(Article().body_nullable, str | None)  # pyrefly: ignore[assert-type] # ty: ignore[type-assertion-failure] # pyright: ignore[reportAssertTypeFailure]
+    assert_type(Article().count_nullable, int | None)  # pyrefly: ignore[assert-type] # ty: ignore[type-assertion-failure] # pyright: ignore[reportAssertTypeFailure]
+    assert_type(Article().slug_nullable, str | None)  # pyrefly: ignore[assert-type] # ty: ignore[type-assertion-failure] # pyright: ignore[reportAssertTypeFailure]
